refactor: remove delete user functionality and introduce disable user service and hook
